pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "md2shunn"
version = "0.9.0"
authors = [
    { name = "Brian Seitz", email = "brnstz@gmail.com" },
]
description = """Converts Markdown (.md) files into Proper Manuscript
Format for fiction writers (a.k.a. Shunn Format, https://www.shunn.net/format/)
as a Word (.docx) file."""
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
license = "MIT"
dependencies = [
    "lxml",
    "marko",
    "python-docx",
]

[project.scripts]
md2shunn = "md2shunn.cli:main"

[project.urls]
"Homepage" = "https://github.com/brnstz/md2shunn"
"Bug Tracker" = "https://github.com/brnstz/md2shunn/issues"
